Fix UI not loading after configuration change (e.g., rotation).
This happened because we only called observe() on LiveData the first time
onCreate() was called, but not subsequently. This is incorrect according to the
docs.

However, to prevent registering multiple identical Observers for the same
Activity instance, we store a reference to the Observer and reuse it if it
already exists.

Quoting from LiveData#observe() docs:

> When it is destroyed, LiveData removes references to the observer & the owner.

And further below, in the context of multiple observe() calls:

> If the given owner, observer tuple is already in the list, the observe() call
> is ignored. If the observer is already in the list with another owner,
> LiveData throws an IllegalArgumentException.